Makes Controller a class with no static state
